% Auto-generated: do not edit by hand
\name{matrix}

\alias{matrix}

\title{Matrix component}

\description{

}

\usage{
matrix(id=NULL, hoverColor=NULL, selectedColor=NULL,
selectedLayer=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{hoverColor}{Character. The color of the layer when hovering over it}

\item{selectedColor}{Character. The color of the layer when selected}

\item{selectedLayer}{Numeric. The layer that needs to be selected}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

